requirements.txt
amqp==5.2.0
annotated-types==0.7.0
APScheduler==3.10.4
babel==2.16.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
billiard==4.2.1
blinker==1.8.2
blis==1.0.1
catalogue==2.0.10
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.19.0
colorama==0.4.6
confection==0.1.5
crontab==1.0.1
cryptography==43.0.3
cymem==2.0.8
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
filelock==3.16.1
Flask==3.0.3
flask-babel==4.0.0
Flask-Bcrypt==1.0.1
Flask-Celery==2.4.3
Flask-Login==0.6.3
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
fsspec==2024.9.0
greenlet==3.1.1
gunicorn==23.0.0
huggingface-hub==0.25.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kombu==5.4.2
langcodes==3.4.1
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.10
mysql-connector-python==9.0.0
networkx==3.4
nltk==3.9.1
numpy==2.0.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
preshed==3.0.9
prompt_toolkit==3.0.48
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.1.1
regex==2024.9.11
requests==2.32.3
rich==13.9.2
safetensors==0.4.5
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
soupsieve==2.6
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.36
srsly==2.4.8
sympy==1.13.3
thinc==8.3.2
tokenizers==0.20.0
torch==2.4.1
tqdm==4.66.5
transformers==4.46.3
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
vine==5.1.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
Werkzeug==3.0.4
wrapt==1.16.0
